♻️ refactor(config): move config page and section logic into domain modules
Move the config route implementation and tests into src/features/config and extract section cache mapping into src/entities/config.

The config store now delegates section extraction to the domain module, making state management narrower.
