+++ src/features/config/ConfigPage.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/mixins' as *;
-@use '../styles/variables' as *;
+@use '../../styles/mixins' as *;
+@use '../../styles/variables' as *;
 
 .container {
   width: 100%;
